Finish average_win_percentage, move teams methods to module
